node: on unrelenting loadavg, restart sooner
when RAM is on the brink and swap is fully used, loadavg steadily starts
shooting up and most of the CPU time is spent in allocs and free. The server
isn't usually able to keep up with incoming requests in this state.

restarting the machine is one avenue to start with a clean slate, at least
until the next time RAM fills up and CPU is tied in swaps and allocs.